:recycle: renderer 파일의 함수 위치 재조정
